Another rule
This adds a rule for the first Canadian property.

It highlights an issue: sometimes there's a city embedded in the street
address, and we don't really have the right context to fix it up here.

Instead, let's defer to a future processing step that has awareness
of things like postal codes and cities.

Next step: refactor the unit-testable pure extraction functions into
their own module and switch to using ECMA imports for the Chrome code.