More rule, test
Try a really aggressive tactic: try to extract every element

Perf sucks, although not as much as you'd expect. We may be able to
apply a heuristic to skip fruitless branches of the DOM - eg if
the element X didn't produce a price field, it's reasonable to
expeect its descendants will be similarly useless and to exclude
them.